Update FMDB and Coreplot, there is still a issue with a non-main thread and Core-Annimation haven’t found it yet but I’m checking in to do an automatic refactoring of core plot to remove GC, this is deprecated but have to wait to release 2.0 to get core plot updated, so trying the auto-refactoring.

+++ Simple Sim/DataController.h
@@ -12,16 +12,7 @@
 @class SignalSystem;
 
 @interface DataController : NSObject{
-//    BOOL _connected;
     BOOL _doThreads;
-//    BOOL cancelProcedure;
-//    BOOL _adhocDataAdded;
-//    id _delegate;
-//    BOOL _fileDataAdded;
-//    NSString *fileDataFileName;
-//    NSArray *fileData;
-//    DataSeries *_dataSeries;
-    
 }
 @property (retain) NSMutableDictionary *archive;
 @property(readonly) BOOL connected;
@@ -73,7 +64,7 @@
                 AndSignalSystem: (SignalSystem *) signalSystem
                     AndDataRate: (long) dataRate
                   WithStoreCode: (long) archiveCode
-       WithRequestTruncatedFlag: (int *) requestTrucated;
+       WithRequestTruncatedFlag: (long *) requestTrucated;
 
 - (long) getMinDataDateTimeForPair:(NSString *) fxPairName;
 - (long) getMaxDataDateTimeForPair:(NSString *) fxPairName;
